requirements.txt
altgraph==0.17
appdirs==1.4.4
argon2-cffi==20.1.0
astroid==2.5.1
async-generator==1.10
attrs==20.3.0
autopep8==1.5.7
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
comtypes==1.1.10
cycler==0.10.0
decorator==5.0.5
defusedxml==0.7.1
distlib==0.3.1
docutils==0.17.1
entrypoints==0.3
ffmpeg==1.4
filelock==3.0.12
future==0.18.2
gTTS==2.2.2
idna==2.10
importlib-metadata==4.5.0
install==1.3.4
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
isort==5.8.0
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
keyring==23.0.1
Kivy==2.0.0
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.sdl2==0.3.1
Kivy-Garden==0.1.4
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
matplotlib==3.4.2
mccabe==0.6.1
mistune==0.8.4
MouseInfo==0.1.3
mysql-connector-python==8.0.24
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.20.2
opencv-python==4.5.2.52
packaging==20.9
pandocfilters==1.4.3
parso==0.8.2
pefile==2021.5.24
pickleshare==0.7.5
Pillow==8.1.2
pkginfo==1.7.0
playsound==1.2.2
prometheus-client==0.10.0
prompt-toolkit==3.0.18
protobuf==3.15.8
PyAutoGUI==0.9.52
pycodestyle==2.7.0
pycparser==2.20
pygame==2.0.1
PyGetWindow==0.0.9
Pygments==2.8.1
pyinstaller-hooks-contrib==2021.1
pylint==2.7.2
pyloco==0.0.139
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
pypiwin32==223
PyRect==0.1.4
pyrsistent==0.17.3
PyScreeze==0.1.26
python-dateutil==2.8.1
pytube==11.0.0
PyTweening==1.0.3
pywin32==300
pywin32-ctypes==0.2.0
pywinpty==0.5.7
pyzmq==22.0.3
readme-renderer==29.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
selenium==3.141.0
Send2Trash==1.5.0
SimpleWebSocketServer==0.1.1
six==1.15.0
soupsieve==2.2.1
terminado==0.9.4
testpath==0.4.4
toml==0.10.2
tornado==6.1
tqdm==4.61.0
traitlets==5.0.5
twine==3.4.1
typing==3.7.4.3
urllib3==1.26.4
ushlex==0.99.1
virtualenv==20.4.4
virtualenvwrapper-win==1.2.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
wikipedia==1.4.0
wrapt==1.12.1
zipp==3.4.1
